take out funds as a separate component in account view
